v7: requires libcore bindings 3.x. drop deprecated code
