+++ src/__tests__/helpers/account.js
@@ -6,9 +6,7 @@ import {
 } from "../../currencies";
 import {
   groupAccountOperationsByDay,
-  groupAccountsOperationsByDay,
-  reorderTokenAccountsByCountervalues,
-  reorderAccountByCountervalues
+  groupAccountsOperationsByDay
 } from "../../account";
 import { genAccount } from "../../mock/account";
 
@@ -45,39 +43,3 @@ test("groupAccountsOperationsByDay", () => {
     flatMap(res1.sections, s => s.data)
   );
 });
-
-test("reorderTokenAccountsByCountervalues", () => {
-  const account = genAccount("toto", {
-    currency: getCryptoCurrencyById("ethereum"),
-    tokenAccountsCount: 10
-  });
-  const tickers = {
-    NEXO: 200000,
-    WEB: 1
-  };
-  const reordered = reorderTokenAccountsByCountervalues(tickers)(
-    account.tokenAccounts || []
-  );
-  expect(reordered.map(ta => ta.token.ticker)).toMatchSnapshot();
-  expect(reorderTokenAccountsByCountervalues(tickers)(reordered)).toBe(
-    reordered
-  );
-});
-
-test("reorderAccountByCountervalues", () => {
-  const account = genAccount("toto2", {
-    currency: getCryptoCurrencyById("ethereum"),
-    tokenAccountsCount: 10
-  });
-  const tickers = {
-    NEXO: 200000,
-    WEB: 1
-  };
-  const reordered = reorderAccountByCountervalues(tickers)(account);
-  expect(
-    (reordered.tokenAccounts || []).map(ta => ta.token.ticker)
-  ).toMatchSnapshot();
-  expect(reorderAccountByCountervalues(tickers)(reordered)).toBe(reordered);
-});
-
-// TODO testing calculateCounterValue is correctly called for picking diff coins/dates.
