+++ src/account/ordering.js
@@ -100,81 +100,3 @@ export const nestedSortAccounts = (
     comparator
   );
 };
-
-// // // // BELOW IS LEGACY // // // //
-
-export type SortAccountsParam = {
-  accounts: Account[],
-  accountsBtcBalance: BigNumber[],
-  orderAccounts: string
-};
-
-type SortMethod = "name" | "balance";
-
-const sortMethod: { [_: SortMethod]: (SortAccountsParam) => string[] } = {
-  balance: ({ accounts, accountsBtcBalance }) =>
-    accounts
-      .map((a, i) => [a.id, accountsBtcBalance[i] || BigNumber(-1), a.name])
-      .sort((a, b) => {
-        const numOrder = a[1].minus(b[1]).toNumber();
-        if (numOrder === 0) {
-          return a[2].localeCompare(b[2]);
-        }
-
-        return numOrder;
-      })
-      .map(o => o[0]),
-
-  name: ({ accounts }) =>
-    accounts
-      .slice(0)
-      .sort((a, b) => a.name.localeCompare(b.name))
-      .map(a => a.id)
-};
-
-export const reorderTokenAccountsByCountervalues = (rates: {
-  [ticker: string]: number // rates by ticker (on the same countervalue reference)
-}) => (tokenAccounts: TokenAccount[]): TokenAccount[] => {
-  const meta = tokenAccounts
-    .map((ta, index) => ({
-      price: ta.balance.times(rates[ta.token.ticker] || 0).toNumber(),
-      ticker: ta.token.ticker,
-      index
-    }))
-    .sort((a, b) => {
-      if (a.price === b.price) {
-        return a.ticker.localeCompare(b.ticker);
-      }
-      return b.price - a.price;
-    });
-  if (meta.every((m, i) => m.index === i)) {
-    // account ordering is preserved, we keep the same array reference (this should happen most of the time)
-    return tokenAccounts;
-  }
-  // otherwise, need to reorder
-  return meta.map(m => tokenAccounts[m.index]);
-};
-
-// high level utility that uses reorderTokenAccountsByCountervalues and keep reference if unchanged
-export const reorderAccountByCountervalues = (rates: {
-  [ticker: string]: number // rates by ticker (on the same countervalue reference)
-}) => (account: Account): Account => {
-  if (!account.tokenAccounts) return account;
-  const tokenAccounts = reorderTokenAccountsByCountervalues(rates)(
-    account.tokenAccounts
-  );
-  if (tokenAccounts === account.tokenAccounts) return account;
-  return { ...account, tokenAccounts };
-};
-
-export function sortAccounts(param: SortAccountsParam) {
-  const [order, sort] = param.orderAccounts.split("|");
-  if (order === "name" || order === "balance") {
-    const ids = sortMethod[order](param);
-    if (sort === "desc") {
-      ids.reverse();
-    }
-    return ids;
-  }
-  return null;
-}
