+++ src/bridge/EthereumJSBridge.js
@@ -586,19 +586,9 @@ export const accountBridge: AccountBridge<Transaction> = {
       };
     }),
 
-  addPendingOperation: (account, operation) => ({
-    ...account,
-    pendingOperations: [operation].concat(
-      account.pendingOperations.filter(
-        o => o.transactionSequenceNumber === operation.transactionSequenceNumber
-      )
-    )
-  }),
-
-  prepareTransaction: (account, transaction) => Promise.resolve(transaction),
-
-  estimateGasLimit: (account, address) => {
-    const api = apiForCurrency(account.currency);
-    return api.estimateGasLimitForERC20(address);
+  prepareTransaction: (a, t) => {
+    const api = apiForCurrency(a.currency);
+    const o = api.estimateGasLimitForERC20(t.recipient);
+    return o.then(gasLimit => ({ ...t, gasLimit: BigNumber(gasLimit) }));
   }
 };
