+++ src/families/ethereum/libcore-buildTransaction.js
@@ -3,7 +3,7 @@
 import invariant from "invariant";
 import { BigNumber } from "bignumber.js";
 import eip55 from "eip55";
-import { FeeNotLoaded } from "@ledgerhq/errors";
+import { FeeNotLoaded, NotEnoughGas, NotEnoughBalance } from "@ledgerhq/errors";
 import type { Account, Transaction } from "../../types";
 import { isValidRecipient } from "../../libcore/isValidRecipient";
 import { bigNumberToLibcoreAmount } from "../../libcore/buildBigNumber";
@@ -68,6 +68,9 @@ export async function ethereumBuildTransaction({
     } else {
       if (!transaction.amount) throw new Error("amount is missing");
       amount = BigNumber(transaction.amount);
+      if (amount.gt(tokenAccount.balance)) {
+        throw new NotEnoughBalance();
+      }
     }
     const to256 = Buffer.concat([
       Buffer.alloc(12),
@@ -116,10 +119,18 @@ export async function ethereumBuildTransaction({
   await transactionBuilder.setGasPrice(gasPriceAmount);
   if (isCancelled()) return;
 
-  const builded = await transactionBuilder.build();
-  if (isCancelled()) return;
+  try {
+    const builded = await transactionBuilder.build();
+    if (isCancelled()) return;
 
-  return builded;
+    return builded;
+  } catch (e) {
+    if (tokenAccount && e.message === "Cannot gather enough funds.") {
+      // FIXME e.message usage: we need a universal error code way. not yet the case with diff bindings
+      throw new NotEnoughGas();
+    }
+    throw e;
+  }
 }
 
 export default ethereumBuildTransaction;
